+++ ui/components/nance/RetroactiveRewards.tsx
@@ -7,7 +7,12 @@ import { useCitizens } from '@/lib/citizen/useCitizen'
 import { useAssets } from '@/lib/dashboard/hooks'
 import { SNAPSHOT_SPACE_NAME } from '@/lib/nance/constants'
 import { useVotingPowers } from '@/lib/snapshot'
-import { iterativeNormalization, minimizeL1Distance } from '@/lib/utils/voting'
+import { getBudget, getPayouts } from '@/lib/utils/rewards'
+import {
+  runIterativeNormalization,
+  getBestFitDistributions,
+  runQuadraticVoting,
+} from '@/lib/utils/voting'
 import Asset from '@/components/dashboard/treasury/balance/Asset'
 import Container from '@/components/layout/Container'
 import ContentLayout from '@/components/layout/ContentLayout'
@@ -40,18 +45,19 @@ export function RetroactiveRewards({
   distributions,
   refreshRewards,
 }: RetroactiveRewardsProps) {
-  const [distribution, setDistribution] = useState<{ [key: string]: number }>(
-    {}
-  )
   const chain =
     process.env.NEXT_PUBLIC_CHAIN === 'mainnet' ? Arbitrum : ArbitrumSepolia
 
-  const year = new Date().getFullYear()
-  const [edit, setEdit] = useState(false)
   const userAddress = useAddress()
   // TODO use current quarter
+  const year = new Date().getFullYear()
   const quarter = Math.floor((new Date().getMonth() + 3) / 3) - 2
+
   // Check if the user already has a distribution for the current quarter
+  const [edit, setEdit] = useState(false)
+  const [distribution, setDistribution] = useState<{ [key: string]: number }>(
+    {}
+  )
   useEffect(() => {
     if (distributions && userAddress) {
       for (const d of distributions) {
@@ -68,17 +74,15 @@ export function RetroactiveRewards({
     }
   }, [userAddress, distributions])
 
-  const projectIdToListOfPercentage: { [key: string]: number[] } = {}
-  const projectIdToEstimatedPercentage: { [key: string]: number } = {}
+  const handleDistributionChange = (projectId: string, value: number) => {
+    setDistribution((prev) => ({
+      ...prev,
+      [projectId]: Math.min(100, Math.max(1, value)),
+    }))
+  }
 
   const addresses = distributions ? distributions.map((d) => d.address) : []
-  const isCitizens = useCitizens(chain, addresses)
-  const citizenDistributions = distributions?.filter((_, i) => isCitizens[i])
-  const nonCitizenDistributions = distributions?.filter(
-    (_, i) => !isCitizens[i]
-  )
 
-  const SNAPSHOT_SPACE_NAME = 'tomoondao.eth'
   const SNAPSHOT_ID =
     '0xa38f7cfeb73b166aea0b65432230bc19faf5411e7f86cc8ea3b961d7c72c85ed'
   const { data: _vps } = useVotingPowers(
@@ -96,82 +100,60 @@ export function RetroactiveRewards({
     addressToQuadraticVotingPower[userAddress] > 0
 
   // All projects need at least one citizen distribution to do iterative normalization
+  const isCitizens = useCitizens(chain, addresses)
+  const citizenDistributions = distributions?.filter((_, i) => isCitizens[i])
+  const nonCitizenDistributions = distributions?.filter(
+    (_, i) => !isCitizens[i]
+  )
   const allProjectsHaveCitizenDistribution = projects.every(({ id }) =>
     citizenDistributions.some(({ distribution }) => id in distribution)
   )
   let votingPowerSumIsNonZero = false
 
+  let projectIdToEstimatedPercentage: { [key: string]: number } = {}
   if (allProjectsHaveCitizenDistribution) {
-    const [filledInCitizenDistributions, votes] = iterativeNormalization(
+    const [filledInCitizenDistributions, votes] = runIterativeNormalization(
       citizenDistributions,
       projects
     )
-    const bestFitNonCitizenDistributions = []
-    for (const d of nonCitizenDistributions) {
-      const desiredDistribution = []
-      const dist = d.distribution
-      for (const [projectIndex, project] of projects.entries()) {
-        desiredDistribution.push(dist[project.id])
-      }
-      const actualDistribution = minimizeL1Distance(desiredDistribution, votes)
-      const newDistribution = {
-        address: d.address,
-        year: d.year,
-        quarter: d.quarter,
-        distribution: actualDistribution,
-      }
-      bestFitNonCitizenDistributions.push(newDistribution)
-    }
+    const bestFitNonCitizenDistributions = getBestFitDistributions(
+      nonCitizenDistributions,
+      projects,
+      votes
+    )
     const allDistributions: Distribution[] = [
       ...filledInCitizenDistributions,
       ...bestFitNonCitizenDistributions,
     ]
-
-    if (distributions) {
-      const allAddresses = allDistributions.map((d) => d.address)
-      for (const d of allDistributions) {
-        const { address, year, quarter, distribution: dist } = d
-        for (const [key, value] of Object.entries(dist)) {
-          if (!projectIdToListOfPercentage[key]) {
-            projectIdToListOfPercentage[key] = []
-          }
-          projectIdToListOfPercentage[key].push(value)
-        }
-      }
-      const votingPowerSum = _.sum(Object.values(addressToQuadraticVotingPower))
-      if (votingPowerSum > 0) {
-        votingPowerSumIsNonZero = true
-        for (const [projectId, percentages] of Object.entries(
-          projectIdToListOfPercentage
-        )) {
-          projectIdToEstimatedPercentage[projectId] =
-            _.sum(
-              percentages.map(
-                (p, i) => p * addressToQuadraticVotingPower[allAddresses[i]]
-              )
-            ) / votingPowerSum
-        }
-        // normalize projectIdToEstimatedPercentage
-        const sum = _.sum(Object.values(projectIdToEstimatedPercentage))
-        for (const [projectId, percentage] of Object.entries(
-          projectIdToEstimatedPercentage
-        )) {
-          projectIdToEstimatedPercentage[projectId] = (percentage / sum) * 100
-        }
-      }
-    }
+    projectIdToEstimatedPercentage = runQuadraticVoting(
+      allDistributions,
+      projects
+    )
   }
 
   const { contract: distributionTableContract } = useContract(
     DISTRIBUTION_TABLE_ADDRESSES[chain.slug]
   )
 
-  const handleDistributionChange = (projectId: string, value: number) => {
-    setDistribution((prev) => ({
-      ...prev,
-      [projectId]: Math.min(100, Math.max(1, value)),
-    }))
-  }
+  const { tokens } = useAssets()
+  const { ethBudget, usdBudget, mooneyBudget, ethPrice } = getBudget(
+    tokens,
+    year,
+    quarter
+  )
+  const {
+    projectIdToETHPayout,
+    projectIdToMooneyPayout,
+    addressToEthPayout,
+    addressToMooneyPayout,
+    ethPayoutCSV,
+    mooneyPayoutCSV,
+  } = getPayouts(
+    projectIdToEstimatedPercentage,
+    projects,
+    ethBudget,
+    mooneyBudget
+  )
 
   const handleSubmit = async () => {
     const totalPercentage = Object.values(distribution).reduce(
@@ -239,74 +221,6 @@ export function RetroactiveRewards({
     }
   }
 
-  const idToTitle: { [key: string]: string } = {}
-  for (const project of projects) {
-    idToTitle[project.id] = project.title
-  }
-
-  const { tokens } = useAssets()
-  const numQuartersPastQ4Y2022 = (year - 2022) * 4 + quarter - 1
-  let ethBudget = 0
-  let usdValue = 0
-  let usdBudget = 0
-  let mooneyBudget = 0
-  let ethPrice = 0
-  if (tokens && tokens[0]) {
-    for (const token of tokens) {
-      if (token.symbol !== 'MOONEY') {
-        usdValue += token.usd
-      }
-    }
-    ethPrice = tokens[0].usd / tokens[0].balance
-    const ethValue = usdValue / ethPrice
-    usdBudget = usdValue * 0.05
-    ethBudget = ethValue * 0.05
-    mooneyBudget = 15_000_000 * 0.95 ** numQuartersPastQ4Y2022
-  }
-
-  const projectIdToETHPayout: { [key: string]: number } = {}
-  const projectIdToMooneyPayout: { [key: string]: number } = {}
-  for (const project of projects) {
-    const percentage = projectIdToEstimatedPercentage[project.id]
-    projectIdToETHPayout[project.id] = (percentage / 100) * ethBudget
-    projectIdToMooneyPayout[project.id] = (percentage / 100) * mooneyBudget
-  }
-  const addressToPayoutProportion: { [key: string]: number } = {}
-  for (const project of projects) {
-    const projectId = project.id
-    const projectPercentage = projectIdToEstimatedPercentage[projectId]
-    const contributors = project.contributors
-    for (const [contributerAddress, contributorPercentage] of Object.entries(
-      contributors
-    )) {
-      if (contributerAddress in addressToPayoutProportion) {
-        addressToPayoutProportion[contributerAddress] +=
-          (contributorPercentage / 100) * (projectPercentage / 100)
-      } else {
-        addressToPayoutProportion[contributerAddress] =
-          (contributorPercentage / 100) * (projectPercentage / 100)
-      }
-    }
-  }
-  const addressToEthPayout: { [key: string]: number } = {}
-  const addressToMooneyPayout: { [key: string]: number } = {}
-  for (const [address, proportion] of Object.entries(
-    addressToPayoutProportion
-  )) {
-    addressToEthPayout[address] = proportion * ethBudget
-    addressToMooneyPayout[address] = proportion * mooneyBudget
-  }
-  const ethPayoutCSV = Object.entries(addressToEthPayout)
-    .map(([address, eth]) => `${address},${eth}`)
-    .join('\n')
-  const mooneyPayoutCSV = Object.entries(addressToMooneyPayout)
-    .map(([address, mooney]) => `${address},${mooney}`)
-    .join('\n')
-  console.log('ethPayoutCSV')
-  console.log(ethPayoutCSV)
-  console.log('mooneyPayoutCSV')
-  console.log(mooneyPayoutCSV)
-
   return (
     <section id="projects-container" className="overflow-hidden">
       <Head title="Projects" image="" />
@@ -324,7 +238,7 @@ export function RetroactiveRewards({
           <section className="w-full flex flex-row items-center sm:items-start">
             <section className="mt-8 flex flex-col w-1/3">
               <h3 className="title-text-colors text-2xl font-GoodTimes">
-                Total Rewards
+                Total Q{quarter} Rewards
               </h3>
               <Asset
                 name="ETH"
